Extracted VisitsFilter type from component for general usage
